+++ Samples~/AR Starter Assets/ARDemoSceneAssets/Scripts/ARSampleMenuManager.cs
@@ -1,10 +1,10 @@
 #if AR_FOUNDATION_PRESENT
+using UnityEngine.Events;
 using UnityEngine.EventSystems;
 using UnityEngine.UI;
 using UnityEngine.XR.Interaction.Toolkit.AR.Inputs;
 using UnityEngine.XR.Interaction.Toolkit.Inputs.Readers;
 using UnityEngine.XR.Interaction.Toolkit.Interactors;
-using UnityEngine.XR.Interaction.Toolkit.Samples.StarterAssets;
 
 namespace UnityEngine.XR.Interaction.Toolkit.Samples.ARStarterAssets
 {
@@ -68,19 +68,6 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.ARStarterAssets
         }
 
         [SerializeField]
-        [Tooltip("The object spawner component in charge of spawning new objects.")]
-        ObjectSpawner m_ObjectSpawner;
-
-        /// <summary>
-        /// The object spawner component in charge of spawning new objects.
-        /// </summary>
-        public ObjectSpawner objectSpawner
-        {
-            get => m_ObjectSpawner;
-            set => m_ObjectSpawner = value;
-        }
-
-        [SerializeField]
         [Tooltip("Button that closes the object creation menu.")]
         Button m_CancelButton;
 
@@ -118,6 +105,20 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.ARStarterAssets
             get => m_TapStartPositionInput;
             set => XRInputReaderUtility.SetInputProperty(ref m_TapStartPositionInput, value, this);
         }
+
+        /// <summary>
+        /// Calls the methods in its invocation list when the spawned object selection changes.
+        /// </summary>
+        public UnityEvent<int> spawnedObjectSelectionChanged
+        {
+            get => m_SpawnedObjectSelectionChanged;
+            set => m_SpawnedObjectSelectionChanged = value;
+        }
+
+        [Header("Events")]
+        [SerializeField]
+        [Tooltip("Calls the methods in its invocation list when the spawned object selection changes.")]
+        UnityEvent<int> m_SpawnedObjectSelectionChanged = new UnityEvent<int>();
 
         bool m_ShowObjectMenu;
 
@@ -173,22 +174,7 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.ARStarterAssets
 
         public void SetObjectToSpawn(int objectIndex)
         {
-            if (m_ObjectSpawner == null)
-            {
-                Debug.LogWarning("Menu Manager not configured correctly: no Object Spawner set.", this);
-            }
-            else
-            {
-                if (objectIndex < m_ObjectSpawner.objectPrefabs.Count)
-                {
-                    m_ObjectSpawner.spawnOptionIndex = objectIndex;
-                }
-                else
-                {
-                    Debug.LogWarning("Object Spawner not configured correctly: object index larger than number of Object Prefabs.", this);
-                }
-            }
-
+            m_SpawnedObjectSelectionChanged.Invoke(objectIndex);
             HideMenu();
         }