+++ Samples~/Spatial Keyboard/Scripts/KeyboardOptimizer.cs
@@ -178,6 +178,7 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.SpatialKeyboard
 
         void GetKeys()
         {
+#if TEXT_MESH_PRO_PRESENT || (UGUI_2_0_PRESENT && UNITY_6000_0_OR_NEWER)
             XRKeyboardKey[] keys = GetComponentsInChildren<XRKeyboardKey>();
             foreach (var keyboardKey in keys)
             {
@@ -189,6 +190,7 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.SpatialKeyboard
                     childPosition = keyboardKey.transform.GetSiblingIndex(),
                 });
             }
+#endif
         }
 
         void GetLayoutGroups()
@@ -217,11 +219,13 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.SpatialKeyboard
                 layoutGroup.enabled = false;
             }
 
+#if TEXT_MESH_PRO_PRESENT || (UGUI_2_0_PRESENT && UNITY_6000_0_OR_NEWER)
             foreach (var keyData in m_KeyData)
             {
                 var key = keyData.key;
                 if (key == null)
                     continue;
+
                 key.transform.SetParent(m_ButtonParentTransform);
 
                 if (key.targetGraphic != null)
@@ -239,6 +243,7 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.SpatialKeyboard
                 if (keyData.batchFollow != null)
                     keyData.batchFollow.enabled = true;
             }
+#endif
         }
 
         /// <summary>
@@ -252,9 +257,10 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.SpatialKeyboard
                 layoutGroup.enabled = true;
             }
 
+#if TEXT_MESH_PRO_PRESENT || (UGUI_2_0_PRESENT && UNITY_6000_0_OR_NEWER)
             foreach (var keyData in m_KeyData)
             {
-                XRKeyboardKey key = keyData.key;
+                var key = keyData.key;
                 if (key == null)
                     continue;
 
@@ -277,11 +283,14 @@ namespace UnityEngine.XR.Interaction.Toolkit.Samples.SpatialKeyboard
                 if (keyData.batchFollow != null)
                     keyData.batchFollow.enabled = false;
             }
+#endif
         }
 
         struct KeyData
         {
+#if TEXT_MESH_PRO_PRESENT || (UGUI_2_0_PRESENT && UNITY_6000_0_OR_NEWER)
             public XRKeyboardKey key;
+#endif
             public KeyboardBatchFollow batchFollow;
             public Transform parent;
             public int childPosition;